ScanMessagesHistoryService
